# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nirmal Connect API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface NewsArticle {
  id: number;
  title: string;
  /** @nullable */
  titleTa?: string | null;
  content: string;
  /** @nullable */
  contentTa?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  featured: boolean;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
}

export interface NewsListResponse {
  items: NewsArticle[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Event {
  id: number;
  title: string;
  /** @nullable */
  titleTa?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionTa?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  venue: string;
  eventDate: string;
  /** @nullable */
  endDate?: string | null;
  category: string;
  createdAt: string;
}

export interface EventListResponse {
  items: Event[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Activity {
  id: number;
  title: string;
  /** @nullable */
  titleTa?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  activityDate: string;
  /** @nullable */
  location?: string | null;
  category: string;
  createdAt: string;
}

export interface ActivityListResponse {
  items: Activity[];
  total: number;
  page: number;
  totalPages: number;
}

export type GalleryItemMediaType =
  (typeof GalleryItemMediaType)[keyof typeof GalleryItemMediaType];

export const GalleryItemMediaType = {
  photo: "photo",
  video: "video",
} as const;

export interface GalleryItem {
  id: number;
  title: string;
  mediaUrl: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  mediaType: GalleryItemMediaType;
  /** @nullable */
  album?: string | null;
  createdAt: string;
}

export interface GalleryListResponse {
  items: GalleryItem[];
  total: number;
  page: number;
  totalPages: number;
}

export interface VolunteerRegistrationBody {
  name: string;
  /** @nullable */
  nameTa?: string | null;
  /** @nullable */
  email?: string | null;
  phone: string;
  /** @nullable */
  ward?: string | null;
  constituency: string;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  message?: string | null;
}

export interface Volunteer {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  ward?: string | null;
  constituency: string;
  status: string;
  createdAt: string;
}

export interface Faq {
  id: number;
  question: string;
  /** @nullable */
  questionTa?: string | null;
  answer: string;
  /** @nullable */
  answerTa?: string | null;
  order: number;
}

export interface ConstituencyStats {
  roadsBuiltKm: number;
  waterProjectsCompleted: number;
  schoolsUpgraded: number;
  healthClinicsOpened: number;
  jobsCreated: number;
  beneficiariesServed: number;
  totalProjects: number;
  completedProjects: number;
  ongoingProjects: number;
}

export interface SiteSummary {
  totalNews: number;
  totalEvents: number;
  totalActivities: number;
  totalGalleryItems: number;
  totalVolunteers: number;
  upcomingEventsCount: number;
}

export type ListNewsParams = {
  page?: number;
  limit?: number;
  featured?: boolean;
};

export type ListEventsParams = {
  page?: number;
  limit?: number;
  upcoming?: boolean;
};

export type ListActivitiesParams = {
  page?: number;
  limit?: number;
};

export type ListGalleryParams = {
  page?: number;
  limit?: number;
  type?: ListGalleryType;
  album?: string;
};

export type ListGalleryType =
  (typeof ListGalleryType)[keyof typeof ListGalleryType];

export const ListGalleryType = {
  photo: "photo",
  video: "video",
} as const;

export type GetFeaturedNewsParams = {
  limit?: number;
};

export type GetUpcomingEventsParams = {
  limit?: number;
};

export type GetRecentActivitiesParams = {
  limit?: number;
};